+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [5.0.0] - 2023-01-27
 ## Added
+- Properties `color` and `swiftUI` allow a SwiftUI `Color` to be obtained from a `TypographyKitColor`.
 
 ## Changed